Added mistral ai support and restructured code
