+++ package.json
@@ -38,6 +38,7 @@
     "@genkit-ai/firebase": "^0.5.3",
     "@genkit-ai/flow": "^0.5.16",
     "@genkit-ai/googleai": "^0.5.16",
+    "@mistralai/mistralai": "^1.3.1",
     "commander": "^11.1.0",
     "dotenv": "^16.4.5",
     "express": "^4.21.0",
